#10 moved api call to a separate java file
