+++ app/src/main/java/com/laika/laika_yellow_book/AsyncResponse.java
@@ -1,0 +1,5 @@
+package com.laika.laika_yellow_book;
+
+public interface AsyncResponse {
+    void processFinish(String output);
+}
